- some unused files removed - HD Wallet can send transactions
