+++ src/App.vue
@@ -42,17 +42,21 @@
             // this.$store.dispatch('Assets/getAllAssets');
             await this.$store.dispatch('Network/init');
 
+
             // check session storage
             // if Remember Keys was enabled, get keys and access wallet
-            this.$store.dispatch('autoAccess').then((res) => {
-                if(res){
-                    parent.$store.dispatch('Notifications/add', {
-                        title: "Keys Remembered",
-                        message: "Your stored keys are used to log you in.",
-                        type: "success"
-                    })
-                }
-            });
+            this.$nextTick(() => {
+                parent.$store.dispatch('autoAccess').then((res) => {
+                    if(res){
+                        parent.$store.dispatch('Notifications/add', {
+                            title: "Keys Remembered",
+                            message: "Your stored keys are used to log you in.",
+                            type: "success"
+                        })
+                    }
+                });
+            })
+
         },
         computed:{
             appReady(){
