+++ src/components/wallet/transfer/TxList.vue
@@ -23,10 +23,11 @@
 
     const uuidv1 = require('uuid/v1');
 
+    import Big from 'big.js';
     import CurrencyInputDropdown from "@/components/misc/CurrencyInputDropdown.vue";
     import AvaAsset from "@/js/AvaAsset";
     import {AssetsDict} from "@/store/modules/assets/types";
-    import {ITransaction} from "@/components/wallet/transfer/types";
+    import {ICurrencyInputDropdownValue, ITransaction} from "@/components/wallet/transfer/types";
 
     @Component({
         components: {
@@ -66,9 +67,15 @@
             this.disabledAssets = res;
         }
 
-        oninputchange(index:number, event:ITransaction):void {
-            this.tx_list[index].asset = event.asset;
-            this.tx_list[index].amount = event.amount;
+        oninputchange(index:number, event:ICurrencyInputDropdownValue):void {
+
+            let asset = event.asset;
+            let amt = event.amount;
+
+            if(!asset) return;
+
+            this.tx_list[index].asset = asset;
+            this.tx_list[index].amount = amt;
 
             this.updateUnavailable();
 
@@ -93,14 +100,14 @@
                 this.tx_list.push({
                     uuid: uuid,
                     asset: this.assets[id],
-                    amount: 0,
+                    amount: Big(0),
                 });
             }else if(this.next_initial){
                 // console.log("initial before: ", this.next_initial);
                 this.tx_list.push({
                     uuid: uuid,
                     asset: this.next_initial,
-                    amount: 0,
+                    amount: Big(0),
                 });
             }
             this.$emit('change', this.tx_list);
