+++ src/components/wallet/transfer/types.ts
@@ -1,8 +1,15 @@
 import AvaAsset from "@/js/AvaAsset";
+import Big from "big.js";
 
 
 export interface ITransaction {
     uuid: string,
     asset: AvaAsset
-    amount: number
+    amount: Big
+}
+
+
+export interface ICurrencyInputDropdownValue {
+    asset: AvaAsset|null,
+    amount: Big
 }
