+++ src/views/Home.vue
@@ -1,5 +1,8 @@
 <template>
     <div class="home">
+        <div v-if="isLoadingPersistKeys">
+            LOADING KEYS
+        </div>
         <div>
             <img class="logo" src="@/assets/wallet_logo.png">
             <router-link to="/access">Access Wallet</router-link>
@@ -7,44 +10,78 @@
         </div>
     </div>
 </template>
-<script>
-    import KeystoreLogin from '@/components/home/KeystoreLogin';
-    import PrivateKeyLogin from "@/components/home/PrivateKeyLogin";
-    import CreateNew from "@/components/home/CreateNew";
+<script lang="ts">
+    import 'reflect-metadata';
+    import { Vue, Component, Prop } from 'vue-property-decorator';
+
+    // import KeystoreLogin from '@/components/home/KeystoreLogin';
+    // import PrivateKeyLogin from "@/components/home/PrivateKeyLogin";
+    // import CreateNew from "@/components/home/CreateNew";
 
 
-    export default {
+    @Component({
         name: 'home',
-        data(){
-            return{
-                loginType: "key", // key || keystore || new
-                privateKey: "",
-                address: "",
-                // If generated...
-                newPrivateKey: "",
-                newPublicKey: "",
-                newAddr: "",
-                menu: 'main', // main | new | access
-            }
-        },
         components: {
-            // QRReader,
-            KeystoreLogin,
-            PrivateKeyLogin,
-            CreateNew
-        },
-        computed: {
-            loginComponent(){
-                if(this.loginType === 'key'){
-                    return PrivateKeyLogin;
-                }else if(this.loginType === 'new'){
-                    return CreateNew;
-                }else{
-                    return KeystoreLogin;
-                }
-            }
+            // KeystoreLogin,
+            // PrivateKeyLogin,
+            // CreateNew
         }
+    })
+    export default class Home extends Vue{
+        loginType:string = "key"; // key || keystore || new
+        privateKey:string = "";
+        address:string = "";
+
+
+        get isLoadingPersistKeys():boolean{
+            return this.$store.state.isLoadingPersistKeys;
+        }
+
+        // get loginComponent(){
+        //     if(this.loginType === 'key'){
+        //         return PrivateKeyLogin;
+        //     }else if(this.loginType === 'new'){
+        //         return CreateNew;
+        //     }else{
+        //         return KeystoreLogin;
+        //     }
+        // }
     }
+
+    // export default {
+    //     name: 'home',
+    //     data(){
+    //         return{
+    //             loginType: "key", // key || keystore || new
+    //             privateKey: "",
+    //             address: "",
+    //             // If generated...
+    //             newPrivateKey: "",
+    //             newPublicKey: "",
+    //             newAddr: "",
+    //             menu: 'main', // main | new | access
+    //         }
+    //     },
+    //     components: {
+    //         // QRReader,
+    //         KeystoreLogin,
+    //         PrivateKeyLogin,
+    //         CreateNew
+    //     },
+    //     computed: {
+    //         isLoadingPersistKeys(){
+    //         },
+    //         loginComponent(){
+    //             if(this.loginType === 'key'){
+    //                 return PrivateKeyLogin;
+    //             }else if(this.loginType === 'new'){
+    //                 return CreateNew;
+    //             }else{
+    //                 return KeystoreLogin;
+    //             }
+    //         }
+    //     }
+    // }
 </script>
 <style scoped lang="scss">
     .home{
